requirements.txt
django
pillow
mysqlclient
mysql-connector-python
django-ckeditor-5
python-decouple
django-environ
whitenoise
pathlib
django-jazzmin
requests
djangorestframework
pyyaml             
pip                    
PyJWT                   
sqlparse               
tzdata                 
urllib3                
yarl    
drf-yasg




